Refactor project structure for Pull Request readiness
- Organized core logic into 'core/' (Google, Todoist, Outlook, Constants).
- Moved automation scripts to 'automation/' (shell script and plist).
- Added '__init__.py' files for proper package management.
- Updated all internal imports and relative paths.
- Cleaned up .gitignore and removed unnecessary backup/IDE files.
