+++ automation/sync_assignments.sh
